move pattern common streaming to internal
