+++ unique_toolkit/tests/experimental/components/streaming/test_pattern_replacer.py
@@ -7,7 +7,7 @@ import re
 import pytest
 
 from unique_toolkit.content.schemas import ContentChunk
-from unique_toolkit.experimental.components.streaming.pattern_replacer import (
+from unique_toolkit.experimental._internal.streaming.pattern_replacer import (
     BATCH_NORMALIZATION_PATTERNS,
     NORMALIZATION_MAX_MATCH_LENGTH,
     NORMALIZATION_PATTERNS,
