+++ unique_toolkit/tests/experimental/integrations/openai/streaming/test_responses_text_delta_event_handler.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 
 import pytest
 
-from unique_toolkit.experimental.components.streaming import TextFlushed
+from unique_toolkit.experimental._internal.streaming import TextFlushed
 from unique_toolkit.experimental.integrations.openai.streaming.event_routing.responses.text_delta_event_handler import (
     ResponsesTextDeltaEventHandler,
 )
