+++ unique_toolkit/unique_toolkit/experimental/_internal/streaming/pattern_replacer.py
