+++ unique_toolkit/unique_toolkit/experimental/integrations/openai/streaming/event_routing/responses/code_interpreter_event_handler.py
@@ -37,7 +37,7 @@ from openai.types.responses.response_code_interpreter_call_interpreting_event im
 )
 
 from unique_toolkit._common.event_bus import TypedEventBus
-from unique_toolkit.experimental.components.streaming import (
+from unique_toolkit.experimental._internal.streaming import (
     ActivityProgressUpdate,
     ActivityStatus,
 )
