+++ unique_toolkit/unique_toolkit/experimental/integrations/openai/streaming/event_routing/responses/text_delta_event_handler.py
@@ -13,8 +13,8 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from unique_toolkit._common.event_bus import TypedEventBus
-from unique_toolkit.experimental.components.streaming import TextFlushed, TextState
-from unique_toolkit.experimental.components.streaming.pattern_replacer import (
+from unique_toolkit.experimental._internal.streaming import TextFlushed, TextState
+from unique_toolkit.experimental._internal.streaming.pattern_replacer import (
     StreamingReplacerProtocol,
 )
