+++ unique_toolkit/unique_toolkit/experimental/integrations/openai/streaming/event_routing/subscribers/message_persister.py
@@ -26,7 +26,7 @@ from typing import TYPE_CHECKING, Any, cast
 
 import unique_sdk
 
-from unique_toolkit.experimental.components.streaming.pattern_replacer import (
+from unique_toolkit.experimental._internal.streaming.pattern_replacer import (
     filter_cited_sdk_references,
 )
